Move pages into docs again, and fix some typo/other
