+++ docs/src/pages/docs/add-components/index.jsx
@@ -6,18 +6,19 @@ import DocsPage from "../_docs-wrapper";
 export default function Docs() {
   return (
     <DocsPage
+      prevUrl={"/docs/quick-start"}
       nextUrl="/docs/plots"
-      middleText="Getting Started"
+      middleText="Add Components"
       prevText="Return to Quick Start"
       nextText="Go to Plots Page"
     >
-      <UsaceBox title="Getting Started">
+      <UsaceBox title="Add Components">
         <H4>Project Setup</H4>
         <div className="gw-mt-3 gw-mb-3">
           <Text>
             Similar to{" "}
             <a href="https://usace.github.io/groundwork">Groundwork</a>,{" "}
-            <b>Groundwork-Water</b>
+            <b>Groundwork-Water </b>
             is a set of React <b>DATA</b> components that are designed to be
             used by USACE Developers building internal and externally facing web
             pages and web apps. If you are just getting started with Groundwork
@@ -44,37 +45,23 @@ export default function Docs() {
             Set up a new project with Vite using the following command:
           </Text>
           <div className="gw-mt-3">
-            <div className="gw-flex gw-flex-row gw-justify-start gw-space-between gw-items-center gw-gap-2 gw-mt-3 gw-mb-3">
-              <Code className="gw-block gw-p-1 gw-px-2" language="bash">
-                {`npm create vite@latest <app-name> --template react`}
-              </Code>
-              <CopyButton text="npm create vite@latest <app-name> --template react" />
-            </div>
+            <Code className="gw-block gw-p-1 gw-px-2" language="bash">
+              {`npm create vite@latest <app-name> --template react`}
+            </Code>
           </div>
         </div>
 
         <H4>Installation</H4>
-        <div className="gw-flex gw-flex-row gw-justify-start gw-space-between gw-items-center gw-gap-2 gw-mt-3 gw-mb-3">
-          <Code className="gw-block gw-p-1 gw-px-2" language="bash">
-            npm install @usace/groundwork
-          </Code>
-          <CopyButton text="npm install @usace/groundwork" />
-        </div>
+        <Code className="gw-block gw-p-1 gw-px-2" language="bash">
+          npm install @usace/groundwork
+        </Code>
         <H4>Import Components and Styles</H4>
-        <div className="gw-flex gw-flex-row gw-justify-start gw-space-between gw-items-center gw-gap-2 gw-mt-3 gw-mb-3">
-          <Code className="gw-block gw-p-1 gw-px-2" language="bash">
-            {`import { SiteWrapper } from "@usace/groundwork"`}
-          </Code>
-          <CopyButton
-            text={`import { SiteWrapper } from "@usace/groundwork"`}
-          />
-        </div>
-        <div className="gw-flex gw-flex-row gw-justify-start gw-space-between gw-items-center gw-gap-2 gw-mt-3 gw-mb-3">
-          <Code className="gw-block gw-p-1 gw-px-2" language="bash">
-            import "@usace/groundwork/dist/style.css"
-          </Code>
-          <CopyButton text={`import "@usace/groundwork/dist/style.css"`} />
-        </div>
+        <Code className="gw-block gw-p-1 gw-px-2" language="bash">
+          {`import { SiteWrapper } from "@usace/groundwork"`}
+        </Code>
+        <Code className="gw-block gw-p-1 gw-px-2" language="bash">
+          import "@usace/groundwork/dist/style.css"
+        </Code>
         <Text>
           Make sure to import style.css from Groundwork into your top-level
           component (i.e. App.jsx), then go build stuff with the components
